fix events

+++ include/objects/StaticObjects/Cookie.h
@@ -11,6 +11,6 @@ public:
     void handleCollision(GameObject& object) { object.handleCollision(*this); }
     void handleCollision(Pacman& player) { player.handleCollision(*this); }
     void handleCollision(Ghost& gameObject) { gameObject.handleCollision(*this); }
-    virtual int checkDistance(const sf::Vector2f& pos){return WINDOW_WIDTH;}
+    virtual float checkDistance(const sf::Vector2f& pos){return WINDOW_WIDTH;}
 
 };
